% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeRepeatSequences-package.R
\docType{package}
\name{analyzeRepeatSequences-package}
\alias{analyzeRepeatSequences}
\alias{analyzeRepeatSequences-package}
\title{analyzeRepeatSequences: Repetitive DNA Sequence Analysis and Visualization}
\description{
Quantitative analysis and data visualization for repetitive DNA sequences.
}
\author{
\strong{Maintainer}: Raquel Moya \email{raquel.alys.moya@gmail.com} (\href{https://orcid.org/0000-0002-0214-7490}{ORCID})

}
\keyword{internal}
