% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_regions.R
\name{calculate_variability_score}
\alias{calculate_variability_score}
\title{Calculate variability score}
\usage{
calculate_variability_score(unitorders_df)
}
\arguments{
\item{unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_example} for format.}
}
\value{
a two-column \code{data.frame} with columns \code{y} (variability scores) and
\code{x} (alignment positions)
}
\description{
Calculates a Shannon's uncertainty score for each position in
the alignment to quantify variability. Variability is computed as:

{\eqn{H(X)=-\sum_{i=1}^{K}p(x)_{i}log_{2}(p(x)_{i})}}

Where:

{\eqn{p(x)_{i}}} is the fraction of repeat units of unit type {\eqn{i}}.

{\eqn{K}} is the number of different repeat units at position {\eqn{X}}.

Gaps are included in the calculation.

Last, the output value is normalized using {\eqn{H(X)_{normalized}=\frac{H(X)}{log_{2}(K)}}}
then a smoothing filter is applied using the R package ksmooth.
}
\examples{
calculate_variability_score(unitorders_example)
}
