% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_sequences.R
\name{call_consensus_sequence}
\alias{call_consensus_sequence}
\title{Define consensus sequence from multiple sequence alignment \code{data.frame}.}
\usage{
call_consensus_sequence(unitorders_df, unit_frequency_df, unit2ascii_df)
}
\arguments{
\item{unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_example} for format.}

\item{unit_frequency_df}{\code{data.frame}cataloging all repeat units observed in
the dataset and their frequency. Has columns \code{seq}, \code{length}, \code{count},
\code{Frequency}, \code{samples},
\code{num_samples}, \code{frac_samples}.
See dataset \code{unit_frequency_example} for format.}

\item{unit2ascii_df}{\code{data.frame} of the conversion
between repeat units and single ASCII characters. Has columns \code{seq},
\code{character}. See dataset \code{unit2ascii} for format.}
}
\value{
a list.
First element: the consensus sequence with alignment gaps.
Second element: the supporting multiple sequence alignment
(rows: individual sequences, columns: alignment position)
}
\description{
\code{call_consensus_sequence} defines a consensus sequence from a multiple
sequence alignment represented by \code{unitorders_df}.
}
\examples{
call_consensus_sequence(unitorders_example, unit_frequency_example, unit2ascii)
}
