% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_msa.R
\name{format_unitorders_to_plot}
\alias{format_unitorders_to_plot}
\title{Format \code{unitorders_df}}
\usage{
format_unitorders_to_plot(unitorders_df, unit2color_df, unitsToHighlight = c())
}
\arguments{
\item{unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_example} for format.}

\item{unit2color_df}{\code{data.frame} of the conversion between repeat
units and colors. Has columns \code{seq}, \code{color}. See dataset \code{unit2color} for
format.}

\item{unitsToHighlight}{Vector of repeat units to show in
color. All other repeat units are shown in beige. Defaults to a zero length
vector.}
}
\value{
a list of length 4 of objects used for plotting by \code{plot_msa}
}
\description{
Formats \code{unitorders_df} for plotting.
}
