% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_regions.R
\name{plot_alleles_at_variable_regions}
\alias{plot_alleles_at_variable_regions}
\title{Plot variable regions}
\usage{
plot_alleles_at_variable_regions(
  allele_at_variable_regions_df,
  unit2ascii_df,
  unit2color_df
)
}
\arguments{
\item{allele_at_variable_regions_df}{\code{data.frame} of a parsed
multiple sequence alignment. Has columns \code{character}, \code{sample}, \code{count},
\code{seq}, \code{gp}. See dataset \code{unitorders_example} for format.}

\item{unit2ascii_df}{\code{data.frame} of the conversion
between repeat units and single ASCII characters. Has columns \code{seq},
\code{character}. See dataset \code{unit2ascii} for format.}

\item{unit2color_df}{\code{data.frame} of the conversion
between repeat units and colors. Has columns \code{seq}, \code{color}. See dataset
\code{unit2color} for format.}
}
\value{
ggplot object showing the alignment of each unique sequence within a
variable region.
}
\description{
For visualizing the unique sequences within variable regions identified by
\code{summarize_variable_regions}.
}
\examples{
plot_alleles_at_variable_regions(vr_alleles_summarized_example,
                                 unit2ascii_df = unit2ascii,
                                 unit2color_df = unit2color)
}
