% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_msa.R
\name{plot_msa}
\alias{plot_msa}
\title{Plot a multiple sequence alignment}
\usage{
plot_msa(
  unitorders_df,
  unit2color_df,
  unitsToHighlight = c(),
  showLegend = F,
  plotSubgroups = F,
  plotVariabilityScore = F,
  plotMinMax = F
)
}
\arguments{
\item{unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_example} for format.}

\item{unit2color_df}{\code{data.frame} of the conversion between repeat
units and colors. Has columns \code{seq}, \code{color}. See dataset \code{unit2color} for
format.}

\item{unitsToHighlight}{Vector of repeat units to show in
color. All other repeat units are shown in beige. Defaults to a zero length
vector.}

\item{showLegend}{\code{logical} indicating whether to plot the legend.
Defaults to FALSE.}

\item{plotSubgroups}{\code{logical} indicating whether \code{unitorders_df} has
column \code{group} for clustering the sequences. Defaults to FALSE. Not
compatible with neither plotVariabilityScore=T nor plotMinMax=T.}

\item{plotVariabilityScore}{\code{logical} indicating whether to plot a score
across the alignment representing repeat unit variability. See
\code{calculate_variability_score} for details. Returns a list of two ggplot objects.}

\item{plotMinMax}{\code{logical} indicating whether to show x-axis
ticks at the first and last repeat units.}
}
\value{
ggplot object showing the alignment of multiple repetitive sequences.
}
\description{
Plots the alignment of multiple repetitive sequences.
}
\examples{
plot_msa(unitorders_example, unit2color)
plot_msa(unitorders_example, unit2color, unitsToHighlight = c("GATCCTGACCTTACTAGTTTACAATCACAG",
                                                              "GACCCTGACCTGACTAGTTTACAATCACAT"))
plot_msa(unitorders_example, unit2color, showLegend = TRUE)
plot_msa(unitorders_withSubgroups_example, unit2color, plotSubgroups = TRUE)
plot_msa(unitorders_example, unit2color, plotVariabilityScore = TRUE)
plot_msa(unitorders_example, unit2color, plotMinMax = TRUE)
}
