% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{positions_list}
\alias{positions_list}
\title{positions_list}
\format{
A vector of sequential integers.
}
\source{
Second element of \code{summarize_variable_regions} output..
}
\usage{
positions_list
}
\description{
Example of multiple sequence alignment positions defining a variable region.
}
\keyword{datasets}
