% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_regions.R
\name{summarize_variable_regions}
\alias{summarize_variable_regions}
\title{Summarize alleles at all variable regions within a multiple sequence alignment}
\usage{
summarize_variable_regions(unitorders_df, threshold)
}
\arguments{
\item{unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_example} for format.}

\item{threshold}{Threshold for variability score (range 0-1) to use for defining variable regions.}
}
\value{
Named list with two elements.
allele_dfs: List of \code{data.frame}, one per variable region. Each row is a unique sequence of a variable region.
positions_lists: List of vectors defining the positions of the multiple sequences alignment that are variable regions.
}
\description{
Defines variable regions by variability score. Formats the input
variable region sequences to \code{summarize_variable_region}.
}
\examples{
summarize_variable_regions(unitorders_example, threshold = 0.25)
}
