% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit2ascii}
\alias{unit2ascii}
\title{unit2ascii}
\format{
A data frame with 37 rows and 2 variables.

\describe{
\item{seq}{Sequence of repeat unit (character)}
\item{character}{Repeat unit encoded as a single ASCII character (character)}
}
}
\source{
Custom .py script to parse repeat units found in dataset.
}
\usage{
unit2ascii
}
\description{
Conversion table for repeat units represented by single characters.
}
\details{
A \code{data.frame} describing the conversion of a repeat unit to a single character
for alignment.
}
\keyword{datasets}
