% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit2color}
\alias{unit2color}
\title{unit2color}
\format{
A data frame with 38 rows and 2 variables.

\describe{
\item{seq}{Sequence of repeat unit (character)}
\item{color}{Hex color (character)}
}
}
\source{
Custom .py script to parse repeat units found in dataset.
}
\usage{
unit2color
}
\description{
Conversion table for repeat units represented by unique color.
}
\details{
A \code{data.frame} describing the color to use for each repeat unit in plots.
Black is used for an alignment gap ("-").
}
\keyword{datasets}
