% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_frequency_example}
\alias{unit_frequency_example}
\title{unit_frequency_example}
\format{
A data frame with 158 rows and 7 variables.

\describe{
\item{seq}{Sequence of repeat unit (character)}
\item{length}{Length in nucleotides of repeat unit (integer)}
\item{count}{Count of repeat unit observed in all sequences (integer)}
\item{Frequency}{Frequency of repeat unit abundance among all repeat units (numeric)}
\item{samples}{Comma-separated list of samples with at least one copy of
repeat unit (character)}
\item{num_samples}{Count of samples with at least one copy of repeat unit (integer)}
\item{frac_samples}{Fraction of total samples with at least one copy of the
repeat unit (numeric)}
}
}
\source{
Custom .py script to parse repeat units found in dataset.
}
\usage{
unit_frequency_example
}
\description{
Table of repeat unit frequencies with no grouping of infrequent repeat units.
}
\details{
A \code{data.frame} describing the frequency of each repeat unit in the dataset.
}
\keyword{datasets}
