% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unitorders_group1_example}
\alias{unitorders_group1_example}
\title{unitorders_group1_example}
\format{
A data frame with 33634 rows and 5 variables.

\describe{
\item{character}{Repeat unit encoded as a single ASCII character (character)}
\item{sample}{Sequence ID (character)}
\item{count}{Width of each repeat unit on the plot (numeric)}
\item{seq}{Sequence of repeat unit (character)}
\item{gp}{Integer that increments one for every row in the ungrouped \code{data.frame}
This gives every repeat unit in a sequence its own position,
rather than clustering identical units together (integer)}
}
}
\source{
Custom .py script to parse FASTA file of multiple sequence alignment.
}
\usage{
unitorders_group1_example
}
\description{
Example of a parsed multiple sequence alignment for one group of sequences.
}
\details{
A \code{data.frame} with the results of a multiple sequence alignment of 134 repetitive sequences.
Each row is one repeat unit. Within a sample, rows are ordered from bottom to top in the 5' to 3' direction.
}
\keyword{datasets}
