% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vr_alleles_summarized_example}
\alias{vr_alleles_summarized_example}
\title{vr_alleles_summarized_example}
\format{
A data frame with 19 rows and 14 variables.

\describe{
\item{allele_num}{Variable region sequence (factor)}
\item{allele}{Variable region sequence (character)}
\item{allele_ascii_rev}{Variable region sequence, reversed (character)}
\item{count}{Number of sequences with variable region sequence (numeric)}
\item{most_common_group}{Group of sequences most commonly identified in (integer)}
\item{samples}{Sequence IDs with variable region sequence (character)}
\item{edit_dist_most_common_allele_nt}{Edit distance in nucleotides to
most common variable region sequence}
\item{edit_dist_second_most_common_allele_nt}{Edit distance in nucleotides to
second most common variable region sequence}
\item{edit_dist_most_common_allele_units}{Edit distance in repeat units to
most common variable region sequence}
\item{edit_dist_second_most_common_allele_units}{Edit distance in repeat units to
second most common variable region sequence}
\item{edit_dist_most_common_allele_units_frequency}{Edit distance in repeat units per
length of variable region sequence}
\item{edit_dist_most_common_allele_nt_frequency}{Edit distance in nucleotides per
length of variable region sequence}
\item{edit_dist_second_most_common_allele_units_frequency}{Edit distance in repeat units per
length of variable region sequence}
\item{edit_dist_second_most_common_allele_nt_frequency}{Edit distance in nucleotides per
length of variable region sequence}
}
}
\source{
First element of \code{summarize_variable_regions} output.
}
\usage{
vr_alleles_summarized_example
}
\description{
Subset of a parsed multiple sequence alignment at identified variable regions.
}
\details{
A \code{data.frame} with counts of the unique sequences found at one variable region.
Each row is one variable region sequence. The sequences can be clustered into
two alleles by edit distances.
}
\keyword{datasets}
