% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_sequences.R
\name{make_repeat_summary_figure}
\alias{make_repeat_summary_figure}
\title{Make summary plot}
\usage{
make_repeat_summary_figure(
  unitorders_group,
  ...,
  unit2color_df,
  unit2ascii_df,
  unit_frequency_df
)
}
\arguments{
\item{unitorders_group}{\code{data.frame} of a parsed multiple sequence
alignment for a group of sequences. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_group1} for format.}

\item{...}{Optional additional \code{unitorders_group} \code{data.frame}s.}

\item{unit2color_df}{\code{data.frame} of the conversion between repeat
units and colors. Has columns \code{seq}, \code{color}. See dataset \code{unit2color} for
format.}

\item{unit2ascii_df}{\code{data.frame} of the conversion
between repeat units and single ASCII characters. Has columns \code{seq},
\code{character}. See dataset \code{unit2ascii} for format.}

\item{unit_frequency_df}{\code{data.frame}cataloging all repeat units observed in
the dataset and their frequency. Has columns \code{seq}, \code{count}, \code{samples}, \code{no_samples}, \code{unit_length}.
See dataset \code{unit_frequency_example} for format.}
}
\value{
ggplot object of several multiple sequence alignments on the same x-axis scale, arranged vertically.
}
\description{
\code{make_repeat_summary_figure} is a plotting function to summarize more than one
multiple sequence alignment and their consensus sequences.
}
\examples{
make_repeat_summary_figure(unitorders_group1_example,
                           unitorders_group2_example,
                           unit2color_df = unit2color,
                           unit2ascii_df = unit2ascii,
                           unit_frequency_df = unit_frequency_example)
}
