% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_msa.R
\name{plot_msa}
\alias{plot_msa}
\title{Plot a multiple sequence alignment}
\usage{
plot_msa(
  unitorders_df,
  unit2ascii_df,
  unit2color_df,
  unitsToHighlight = c(),
  showLegend = F,
  plotSubgroups = F,
  plotVariabilityScore = F,
  plotVariableRegion = F,
  plotMinMax = F
)
}
\arguments{
\item{unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment. Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See
dataset \code{unitorders_example} for format.}

\item{unit2ascii_df}{\code{data.frame} of the conversion between repeat
units and single ASCII characters. Has columns \code{seq}, \code{character}. See
dataset \code{unit2ascii} for format.}

\item{unit2color_df}{\code{data.frame} of the conversion between repeat
units and colors. Has columns \code{seq}, \code{color}. See dataset \code{unit2color} for
format.}

\item{unitsToHighlight}{Vector of selected repeat units to show in
color. All other repeat units are shown in beige. Defaults to zero length
vector.}

\item{showLegend}{\code{logical} indicating whether to plot the legend. Defaults to FALSE.}

\item{plotSubgroups}{\code{logical} indicating whether \code{unitorders_df} has
column \code{group} for clustering the sequences. Defaults to FALSE.}

\item{plotVariableRegion}{\code{logical} indicating}

\item{plotMinMax}{\code{logical} indicating}
}
\description{
\code{plot_msa} returns a ggplot object showing the alignment of multiple repetitive sequences.
}
