% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_regions.R
\name{summarize_variable_region}
\alias{summarize_variable_region}
\title{Summarize alleles at variable region}
\usage{
summarize_variable_region(vr_unitorders_df, positions_list)
}
\arguments{
\item{vr_unitorders_df}{\code{data.frame} of a parsed multiple sequence
alignment within the positions defined as variable regions by \code{summarize_variable_regions}.
Has columns \code{character}, \code{sample}, \code{count}, \code{seq}, \code{gp}. See dataset \code{unitorders_example} for format.}
}
\value{
a two-column \code{data.frame} with columns \code{y} (variability scores) and
\code{x} (alignment positions)
}
\description{
Defines the most common sequence and the second most common sequence
as two alleles within a variable region. Calculates the edit distance between these
alleles the other sequences within the variable region. With this, users can use
edit distance thresholds to assign each sequence to an allele.

Note: This function defines exactly two alleles. There may be more
alleles that have high edit distances to both.
}
\examples{
calculate_variability_score(unitorders_example)
}
